+++ elements/listable.go
@@ -1,0 +1,6 @@
+package elements
+
+type Listable interface {
+	GetLevel()
+	GetName()
+}
