--- ui/outputs/layers.go
@@ -1,2 +1,0 @@
-package outputs
-
