+++ ui/outputs/painting.go
@@ -8,7 +8,7 @@ import (
 
 	g "github.com/AllenDang/giu"
 	// "github.com/AllenDang/imgui-go"
-	"github.com/LuckyG0ldfish/GraphicalGo/elements"
+	// "github.com/LuckyG0ldfish/GraphicalGo/elements"
 	"github.com/LuckyG0ldfish/GraphicalGo/elements/subelements"
 )
 
@@ -23,7 +23,6 @@ func DrawUI() {
 			g.Custom(func() {
 				canvas := g.GetCanvas()
 				AddOverView(canvas)
-				// AddObjectSelect(canvas)
 
 				for _, v := range Project.Can.Dragables {
 					AddDragable(canvas, v)
@@ -32,75 +31,10 @@ func DrawUI() {
 				for _, v := range Project.Obj.Pressables {
 					AddButton(canvas, v)
 				}
-
-				for _, v := range Project.Over.Pressables {
-					AddButton(canvas, v)
-				}
 			}),
 		// ),
 		
 	)
-}
-
-func AddDragable(c *g.Canvas, drag elements.Dragable) {
-	typeID := drag.GetType()
-	switch typeID {
-	case 1:	// ID Folder 
-		AddFolder(c, drag)
-	case 2:	// ID File 
-		AddFile(c, drag)
-	case 3:	// ID Object 
-		AddObject(c, drag)
-	} 
-}
-
-func AddFolder(c *g.Canvas, drag elements.Dragable) {
-	pos := g.GetCursorScreenPos()
-
-	mix := drag.GetXLeft()
-	miy := drag.GetYTop()
-
-	max := drag.GetXRight()
-	may := drag.GetYBot()
-
-	c.AddRectFilled(pos.Add(image.Pt(mix, miy)), pos.Add(image.Pt(mix+110, miy+20)), color.White, 0, 5)
-	c.AddRectFilled(pos.Add(image.Pt(mix, miy+20)), pos.Add(image.Pt(max, may)), color.White, 0, 5)
-	c.AddLine(pos.Add(image.Pt(mix+3, miy+20)), pos.Add(image.Pt(mix+107, miy+20)), color.Black, 1)
-	c.AddText(pos.Add(image.Pt(mix+3, miy+3)), color.Black, drag.GetName())
-
-	// expand
-	c.AddRectFilled(pos.Add(image.Pt(max-6, may-6)), pos.Add(image.Pt(max-1, may-1)), color.Black, 0, 5)
-}
-
-func AddFile(c *g.Canvas, drag elements.Dragable) {
-	
-}
-
-func AddObject(c *g.Canvas, drag elements.Dragable) {
-	pos := g.GetCursorScreenPos()
-
-	mix := drag.GetXLeft()
-	miy := drag.GetYTop()
-
-	max := drag.GetXRight()
-	may := drag.GetYBot()
-
-	c.AddRectFilled(pos.Add(image.Pt(mix, miy)), pos.Add(image.Pt(max, may)), color.White, 0, 5)
-	c.AddLine(pos.Add(image.Pt(mix+3, miy+20)), pos.Add(image.Pt(max-3, miy+20)), color.Black, 1)
-	c.AddText(pos.Add(image.Pt(mix+3, miy+3)), color.Black, drag.GetName())
-}
-
-func AddButton(c *g.Canvas, pres elements.Pressable) {
-	pos := g.GetCursorScreenPos()
-	
-	mix := pres.GetXLeft()
-	miy := pres.GetYTop()
-
-	max := pres.GetXRight()
-	may := pres.GetYBot()
-
-	c.AddRectFilled(pos.Add(image.Pt(mix, miy)), pos.Add(image.Pt(max, may)), color.White, 0, 5)
-	c.AddText(pos.Add(image.Pt(mix+3, miy+3)), color.Black, pres.GetName())
 }
 
 func AddSeperation(c *g.Canvas, XCord int) {
@@ -109,14 +43,4 @@ func AddSeperation(c *g.Canvas, XCord int) {
 	c.AddLine(pos.Add(image.Pt(XCord, 0)), pos.Add(image.Pt(XCord, Project.Win.YHeight)), color.White, 1)
 }
 
-func AddOverView(c *g.Canvas) {
-	Project := subelements.GetPro()
-	pos := g.GetCursorScreenPos()
-	c.AddRectFilled(pos.Add(image.Pt(-20, -20)), pos.Add(image.Pt(Project.Can.XLeft, Project.Win.YHeight)), color.White, 0, 5)
-}
 
-func AddObjectSelect(c *g.Canvas) {
-	Project := subelements.GetPro()
-	pos := g.GetCursorScreenPos()
-	c.AddRectFilled(pos.Add(image.Pt(Project.Win.XWidth, -20)), pos.Add(image.Pt(Project.Can.XRight, Project.Win.YHeight)), color.White, 0, 5)
-}
